Insert all tracks to local db

+++ app/src/main/java/com/ashehata/me_player/modules/home/data/model/TrackDataModel.kt
@@ -10,6 +10,6 @@ data class TrackDataModel(
     val id: Long? = null,
     val name: String,
     val uri: String,
-    val duration: Long,
-    val size: Long,
+    val duration: Int,
+    val size: Int,
 )
